refactor: use context getter hook
