+++ src/features/MainScreen/Notifications/Notifications.tsx
@@ -5,8 +5,7 @@ import { FaBell, FaXmark } from 'react-icons/fa6';
 import { cn } from '@bem-react/classname';
 import { Icon } from '@components/Icon/Icon.bundle/common';
 import { Stack } from '@components/Stack/Stack';
-
-import { useBottomPanelManager } from '../StatusBar';
+import { useBottomPanelManager } from '@features/MainScreen/StatusBar/StatusBarProvider';
 
 import './Notifications.css';
