+++ src/features/MainScreen/StatusBar/index.tsx
@@ -1,45 +1,14 @@
-import React, {
-	createContext,
-	FC,
-	HTMLProps,
-	PropsWithChildren,
-	useContext,
-} from 'react';
+import React, { FC, HTMLProps } from 'react';
 import { Button } from 'react-elegant-ui/esm/components/Button/Button.bundle/desktop';
-import { ButtonsManagerObject, useButtonsManager } from '@api/buttons/useButtonsManager';
 import { cn } from '@bem-react/classname';
 import { Icon } from '@components/Icon/Icon.bundle/common';
 import { Stack } from '@components/Stack/Stack';
 import { INotesController } from '@core/features/notes/controller';
+import { useStatusBar } from '@features/MainScreen/StatusBar/StatusBarProvider';
 
 import './StatusBar.css';
 
-export const BottomPanelManagerContext = createContext<ButtonsManagerObject>(
-	null as unknown as ButtonsManagerObject,
-);
-
-export const useBottomPanelManager = () => {
-	return useContext(BottomPanelManagerContext);
-};
-
 export const cnStatusBar = cn('StatusBar');
-
-export const StatusBarContext = createContext<ButtonsManagerObject['state']>({
-	start: [],
-	end: [],
-});
-
-export const StatusBarProvider: FC<PropsWithChildren> = ({ children }) => {
-	const buttonsManager = useButtonsManager();
-
-	return (
-		<StatusBarContext.Provider value={buttonsManager.state}>
-			<BottomPanelManagerContext.Provider value={buttonsManager}>
-				{children}
-			</BottomPanelManagerContext.Provider>
-		</StatusBarContext.Provider>
-	);
-};
 
 export type StatusBarProps = HTMLProps<HTMLDivElement> & {
 	notesRegistry: INotesController;
@@ -53,7 +22,7 @@ export const StatusBar: FC<StatusBarProps> = ({
 	updateNotes,
 	...props
 }) => {
-	const { start, end } = useContext(StatusBarContext);
+	const { start, end } = useStatusBar();
 
 	return (
 		<div {...props} className={cnStatusBar({}, [className])}>
